requirements/base.txt
# basics
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
django-extensions==3.2.1
django-storages==1.13.2
drf-extra-fields==3.4.1
psycopg2==2.9.5
channels==4.0.0
daphne==4.0.0
pillow==9.4.0
boto3==1.26.75
drf-yasg==1.21.5
redis==4.5.1
celery==5.2.7
bleach==6.0.0
